requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djoser==2.2.0
psycopg2-binary==2.9.6
python-environ==0.4.54
drf-yasg==1.21.6
flake8==6.0.0
django_debug_toolbar==3.8.1
Pillow==9.5.0